Correcao da retrocompatibilidade, voltando a classe FlatFileReaderDefinition e deletando a classe FaltFileReaderMultiDefinition
