+++ ffpojo-samples/src/main/java/com/github/ffpojo/samples/SimpleFileSystemFlatFileReaderWithDefaultFlatFileProcessorExample.java
@@ -10,7 +10,7 @@ import com.github.ffpojo.file.processor.record.RecordProcessor;
 import com.github.ffpojo.file.processor.record.event.RecordEvent;
 import com.github.ffpojo.file.reader.FileSystemFlatFileReader;
 import com.github.ffpojo.file.reader.FlatFileReader;
-import com.github.ffpojo.file.reader.FlatFileReaderMultiDefinition;
+import com.github.ffpojo.file.reader.FlatFileReaderDefinition;
 import com.github.ffpojo.metadata.positional.annotation.PositionalField;
 import com.github.ffpojo.metadata.positional.annotation.PositionalRecord;
 
@@ -92,7 +92,7 @@ public class SimpleFileSystemFlatFileReaderWithDefaultFlatFileProcessorExample {
 		if (!inputFile.exists()) {
 			throw new IllegalStateException("File not found: " + INPUT_TXT_OS_PATH);
 		}
-		FlatFileReaderMultiDefinition ffDefinition = new FlatFileReaderMultiDefinition(Customer.class);
+		FlatFileReaderDefinition ffDefinition = new FlatFileReaderDefinition(Customer.class);
 		FlatFileReader ffReader = new FileSystemFlatFileReader(inputFile, ffDefinition);
 		FlatFileProcessor ffProcessor = new DefaultFlatFileProcessor(ffReader);
 		ffProcessor.processFlatFile(new CustomerRecordProcessor());
