+++ ffpojo-samples/src/main/java/com/github/ffpojo/samples/SimpleFileSystemFlatFileReaderWithThreadPoolFlatFileProcessorExample.java
@@ -10,7 +10,7 @@ import com.github.ffpojo.file.processor.record.RecordProcessor;
 import com.github.ffpojo.file.processor.record.event.RecordEvent;
 import com.github.ffpojo.file.reader.FileSystemFlatFileReader;
 import com.github.ffpojo.file.reader.FlatFileReader;
-import com.github.ffpojo.file.reader.FlatFileReaderMultiDefinition;
+import com.github.ffpojo.file.reader.FlatFileReaderDefinition;
 import com.github.ffpojo.metadata.positional.annotation.PositionalField;
 import com.github.ffpojo.metadata.positional.annotation.PositionalRecord;
 
@@ -95,7 +95,7 @@ public class SimpleFileSystemFlatFileReaderWithThreadPoolFlatFileProcessorExampl
 		}
 		
 		
-		FlatFileReaderMultiDefinition ffDefinition = new FlatFileReaderMultiDefinition(Customer.class);
+		FlatFileReaderDefinition ffDefinition = new FlatFileReaderDefinition(Customer.class);
 		FlatFileReader ffReader = new FileSystemFlatFileReader(inputFile, ffDefinition);
 		FlatFileProcessor ffProcessor = new ThreadPoolFlatFileProcessor(ffReader, 5);
 		ffProcessor.processFlatFile(new CustomerRecordProcessor());
