+++ ffpojo-samples/src/main/resources/FileSystemFlatFileReaderWithHeaderAndTrailerExample.txt
@@ -1,5 +1,6 @@
 654035689810/10/2010
-25654Gilberto Holms      gibaholms@hotmail.com         
+25654Gilberto Holms      gibaholms@hotmail.com
+12345William Miranda	 blackstile@hotmail.com         
 65987Billy Idol          billy@thehost.com             
 25465Axel Rose           axl@thehost.com               
 98653Sebastian Bach      sebastian@thehost.com
