+++ ffpojo-samples/src/main/resources/PositionalRecordParsingWithDecoratorExampleInput.txt
@@ -1,4 +1,5 @@
 25654Gilberto Holms      gibaholms@hotmail.com         21/05/1985
+25654William Miranda     blackstile@hotmail.com        08/07/1980
 65987Billy Idol          billy@thehost.com             10/02/1920
 25465Axel Rose           axl@thehost.com               14/11/2001
 98653Sebastian Bach      sebastian@thehost.com         16/10/1960
