+++ ffpojo-samples/src/main/resources/SimpleDelimitedRecordParsingExampleInput.txt
@@ -1,4 +1,5 @@
 25654|Gilberto Holms|gibaholms@hotmail.com
+12345|William Miranda|blackstile@hotmail.com
 65987|Billy Idol|billy@thehost.com
 25465|Axel Rose|axl@thehost.com
 98653|Sebastian Bach|sebastian@thehost.com
