+++ ffpojo-samples/src/main/resources/SimplePositionalRecordParsingExampleInput.txt
@@ -1,4 +1,5 @@
 25654Gilberto Holms      gibaholms@hotmail.com         
+12345William Miranda     blackstile@hotmail.com         
 65987Billy Idol          billy@thehost.com             
 25465Axel Rose           axl@thehost.com               
 98653Sebastian Bach      sebastian@thehost.com
